+++ test/b.smoke.test.js
@@ -1,15 +1,15 @@
 const _ = require('lodash')
-const { test_data_b } = require('../b')
-
+const { mock_b } = require('../b')
+const { mocker } = require('../test_helper')
 //
 describe('mock test', () => {
   //
-  test.each(test_data_b)('本尊與 mock 執行結果應相同', ({ m, n }) => {
-    const { b, mocked_b } = require('../b')
+  test.each(mock_b)('本尊與 mock 執行結果應相同', ({ m, n }) => {
+    const { b } = require('../b')
 
     // 同組測資分別餵給本尊與分身各跑一次然後比較兩者答案是否相同
     const truth = b(m, n)
-    const fake = mocked_b(test_data_b)(m, n)
+    const fake = mocker(mock_b)(m, n)
     // console.debug(`比較結果`, { truth, fake })
 
     // 驗証
@@ -22,7 +22,7 @@ describe('mock test', () => {
   //
   test('傳入不存在的測資參數應失敗', () => {
     const { mocked_b } = require('../b')
-    const result = mocked_b(test_data_b)(0, 1)
+    const result = mocker(mock_b)(0, 1)
     // console.debug(`測試結果:`, result)
 
     // 驗証