+++ test/b.unit.test.js
@@ -1,11 +1,10 @@
 const _ = require('lodash')
-// 拿已建好的相同測資來驗証 b 計算結果是否正確
-const { test_data_b } = require('../b')
+const { mock_b } = require('../b')
 
 //
 describe('b unit test', () => {
   // 用預先寫好的測資與答案去驗算 b 計算結果是否正確
-  test.each(test_data_b)('b 各種計算應成功', ({ m, n, ans }) => {
+  test.each(mock_b)('b 各種計算應成功', ({ m, n, ans }) => {
     const { b } = require(`../b`)
     const result = b(m, n)
     // console.debug(`計算結果`, {result, ans, m, n} )